extract analytics interface
